requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
sqlalchemy==2.0.38
asyncmy==0.2.11
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
PyPDF2==4.0.1
python-docx==0.8.11
